requirements.txt
pyspark
kafka-python
pymongo==4.4.0